get cljs code running shadow-cljs
