+++ entry/index.html
@@ -1,2 +1,2 @@
 
-<html><head><title>Quatrefoil</title><link href="https://avatars1.githubusercontent.com/u/10102840?v=3&s=200" rel="icon" type="image/png"></link><link></link><meta charset="utf-8"></meta><meta content="width=device-width, initial-scale=1" name="viewport"></meta><meta content="{}" id="ssr-stages"></meta><style>body {margin: 0;}</style><style>body * {box-sizing: border-box;}</style><script id="config" type="text/edn">{:build? true}</script></head><body><div id="app"></div><script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/84/three.min.js"></script><script src="object3d.js"></script><script src="main.js"></script></body></html>
+<html><head><title>Quatrefoil</title><link href="https://avatars1.githubusercontent.com/u/10102840?v=3&s=200" rel="icon" type="image/png"></link><link></link><meta charset="utf-8"></meta><meta content="width=device-width, initial-scale=1" name="viewport"></meta><meta content="{}" id="ssr-stages"></meta><style>body {margin: 0;}</style><style>body * {box-sizing: border-box;}</style><script id="config" type="text/edn">{:build? true}</script></head><body><canvas id="canvas"></canvas><script src="client.js"></script></body></html>
