+++ js/quatrefoil/deps.cljs
@@ -1,0 +1,2 @@
+
+{:npm-deps {"three" "^0.114.0"}}
