--- src/quatrefoil/comp/container.cljs
@@ -1,18 +1,0 @@
-
-(ns quatrefoil.comp.container
-  (:require [hsl.core :refer [hsl]]
-            [respo-ui.style :as ui]
-            [respo.alias :refer [create-comp div span canvas]]
-            [respo.comp.space :refer [comp-space]]
-            [respo.comp.text :refer [comp-text]]))
-
-(def comp-container
-  (create-comp
-   :container
-   (fn [store]
-     (fn [state mutate!]
-       (div
-        {:style (merge ui/global)}
-        (canvas
-         {:attrs {:id "canvas", :width js/window.innerWidth, :height js/window.innerHeight},
-          :style {}}))))))
