+++ src/quatrefoil/dsl/alias.cljs
@@ -23,24 +23,24 @@
    (arrange-children children)
    nil))
 
-(defn point-light [props & children] (create-element :point-light props children))
-
-(defn perspective-camera [props & children]
-  (create-element :perspective-camera props children))
-
-(defn group [props & children] (create-element :group props children))
+(defn box [props & children] (create-element :box props children))
 
 (defn camera [props & children] (create-element :camera props children))
 
 (defn create-comp [comp-name hooks render]
   (fn [& args] (Component. comp-name args {} {} render nil hooks false)))
 
-(defn box [props & children] (create-element :box props children))
-
-(defn text [props & children] (create-element :text props children))
+(defn group [props & children] (create-element :group props children))
 
 (defn line [props & children] (create-element :line props children))
 
-(defn sphere [props & children] (create-element :sphere props children))
+(defn perspective-camera [props & children]
+  (create-element :perspective-camera props children))
+
+(defn point-light [props & children] (create-element :point-light props children))
 
 (defn scene [props & children] (create-element :scene props children))
+
+(defn sphere [props & children] (create-element :sphere props children))
+
+(defn text [props & children] (create-element :text props children))
